feat: vad
Java调用Silero VAD 模型很麻烦

+++ main/xiaozhi-server-java/src/main/java/com/ephemetra/server/model/props/SileroVAD.java
@@ -8,8 +8,8 @@ import lombok.Data;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class SileroVAD {
     public String type;
-    public Double threshold;
-    public Double thresholdLow;
+    public Float threshold;
+    public Float thresholdLow;
     public String modelDir;
     public Integer minSilenceDurationMs;
 }
